add server manager and move code
